Fixed the error when trying to convert the object of type 'System.Int32' to the type 'System.Int16'.

+++ Hybrsoft.EnterpriseManager/Converters/Int16Converter.cs
@@ -31,7 +31,7 @@ namespace Hybrsoft.EnterpriseManager.Converters
 					return n16;
 				}
 			}
-			return 0;
+			return (Int16)0;
 		}
 	}
 }
